Introduce new features

+++ src/components/Products/DashboardCard/index.tsx
@@ -1,10 +1,13 @@
-import Button from "components/utils/Button";
+import { useState } from "react";
 import Image from "next/image";
-import React, { useState } from "react";
-import { MdDelete, MdModeEdit } from "react-icons/md";
+import Link from "next/link";
+
 import { useQueryClient } from "react-query";
 import { toast } from "react-toastify";
 import { formatPrice } from "services/format";
+
+import Button from "components/utils/Button";
+import { MdDelete, MdModeEdit } from "react-icons/md";
 import { deleteProduct, updateProductAvailableById } from "utils/http";
 
 import * as S from "./styles";
@@ -12,11 +15,12 @@ import * as S from "./styles";
 import imageFallback from "/public/noimageavailable.svg";
 
 type DashboardCardProps = {
+  companyId: string;
   id: string;
   title: string;
   price: number;
   description: string;
-  available: number;
+  availableSalle: number;
   image: string;
 };
 
@@ -25,18 +29,24 @@ const DashboardCard: React.FC<DashboardCardProps> = ({
   title,
   price,
   description,
-  available,
+  availableSalle,
   image,
+  companyId,
 }: DashboardCardProps) => {
-  const availableItem = available === 1 ? true : false;
+  const [isAvailable, setIsAvailable] = useState(availableSalle);
 
-  const [isAvailable, setIsAvailable] = useState(availableItem);
   const queryClient = useQueryClient();
 
-  function handleDeleteProduct() {
-    try {
+  async function DeleteProduct(id: string) {
+    const confirmation = confirm(
+      "Você tem certeza que deseja excluir este produto?"
+    );
+
+    if (confirmation) {
       deleteProduct(id);
-      queryClient.invalidateQueries("products");
+
+      await queryClient.refetchQueries();
+
       toast.success("Produto deletado com sucesso! 🙂", {
         position: "top-right",
         autoClose: 5000,
@@ -46,7 +56,37 @@ const DashboardCard: React.FC<DashboardCardProps> = ({
         draggable: true,
         progress: undefined,
       });
-    } catch (e) {
+    } else {
+      console.log("Cancelado");
+    }
+  }
+
+  async function updateStatusAvailableProduct(id: string) {
+    if (isAvailable === 1) {
+      setIsAvailable(0);
+      await updateProductAvailableById(id, 0);
+    } else if (isAvailable === 0) {
+      setIsAvailable(1);
+      await updateProductAvailableById(id, 1);
+    }
+
+    toast.success("Produto atualizado com sucesso! 🙂", {
+      position: "top-right",
+      autoClose: 5000,
+      hideProgressBar: false,
+      closeOnClick: true,
+      pauseOnHover: true,
+      draggable: true,
+      progress: undefined,
+    });
+  }
+
+  async function HandleDeleteProduct() {
+    try {
+      await DeleteProduct(id);
+
+      await queryClient.refetchQueries();
+    } catch (error) {
       toast.error("Não foi possivel deletar o produto! 😢", {
         position: "top-right",
         autoClose: 5000,
@@ -59,22 +99,9 @@ const DashboardCard: React.FC<DashboardCardProps> = ({
     }
   }
 
-  function handleChangeProductAvailable() {
+  async function handleChangeProductAvailable() {
     try {
-      setIsAvailable((prevState) => !prevState);
-      console.log(isAvailable);
-      updateProductAvailableById(id, isAvailable === true ? 1 : 0);
-
-      queryClient.invalidateQueries("products");
-      toast.success("Produto atualizado com sucesso! 🙂", {
-        position: "top-right",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-      });
+      updateStatusAvailableProduct(id);
     } catch (e) {
       toast.error("Não foi possivel atualizar o produto! 😢", {
         position: "top-right",
@@ -106,16 +133,24 @@ const DashboardCard: React.FC<DashboardCardProps> = ({
       <S.Description>{description}</S.Description>
       <S.Value>{formatPrice(price)}</S.Value>
       <S.Available>
-        <S.Title>{isAvailable ? "Disponivel" : "Indisponivel"}</S.Title>
+        <S.Title>{isAvailable === 1 ? "Disponivel" : "Indisponivel"}</S.Title>
         <Button onClick={handleChangeProductAvailable}>
-          {isAvailable ? "Deixar Indisponivel" : "Deixar Disponivel"}
+          {isAvailable === 1 ? "Deixar Indisponivel" : "Deixar Disponivel"}
         </Button>
       </S.Available>
       <S.ButtonWrapper>
-        <S.UpdateProduct>
-          <MdModeEdit size="25px" />
-        </S.UpdateProduct>
-        <S.RemoveProduct onClick={handleDeleteProduct}>
+        <Link
+          href={{
+            pathname: "/[companyId]/home/produtos/atualizar-produto/[id]",
+            query: { companyId, id },
+          }}
+        >
+          <S.UpdateProduct>
+            <MdModeEdit size="25px" />
+          </S.UpdateProduct>
+        </Link>
+
+        <S.RemoveProduct onClick={HandleDeleteProduct}>
           <MdDelete size="25px" />
         </S.RemoveProduct>
       </S.ButtonWrapper>
